+++ frontend/src/components/Didits/Didit.jsx
@@ -1,7 +1,6 @@
 import { format } from "date-fns";
 
 const Didit = ({ title, completedOn, project }) => {
-  console.log(project);
   return (
     <div className="flex justify-between px-40 ">
       <div className="flex gap-3 ">
