+++ frontend/src/components/Tasks/TaskForm.jsx
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import DueDate from "../DueDate";
+import DueDate from "./DueDate";
 import Button from "../UI/Button";
 import Input from "../UI/Input";
